+++ tianshou/trainer/base.py
@@ -562,16 +562,17 @@ class Trainer(Generic[TAlgorithm, TTrainerParams], ABC):
                 t.update(training_step_result.get_steps_in_epoch_advancement())
                 self._stop_fn_flag = training_step_result.is_training_done()
                 self._env_step += training_step_result.get_env_step_advancement()
+                training_stats = training_step_result.get_training_stats()
+                assert training_stats is not None
                 TraceLogger.log(
                     log,
-                    lambda: f"Training step complete: stats={training_step_result.get_training_stats().get_loss_stats_dict()}",
+                    lambda: f"Training step complete: stats={training_stats.get_loss_stats_dict()}",
                 )
                 self._log_params(self.algorithm)
 
                 collect_stats = training_step_result.get_collect_stats()
                 if collect_stats is not None:
                     self._logger.log_train_data(asdict(collect_stats), self._env_step)
-                training_stats = training_step_result.get_training_stats()
 
                 pbar_data_dict = self._create_epoch_pbar_data_dict(training_step_result)
                 pbar_data_dict = set_numerical_fields_to_precision(pbar_data_dict)